requirements.txt
# PlastidTaxa Requirements (Refactored)

# Core scientific computing
numpy>=2.0
scipy>=1.15
pandas>=2.0

# Visualization
matplotlib>=3.8
seaborn>=0.13
plotly>=6.0

# Bioinformatics - Core
# NOTE: Biopython >=1.87 patches XXE vulnerability in Bio.Entrez
biopython>=1.87

# XXE Mitigation - Safe XML parsing for Entrez & other external sources
defusedxml>=0.0.1

# DADA2 alternative (pure Python implementation)
# Fallback: can use R/rpy2 if needed, but pure Python is preferred
# dada2>=1.30  # Uncomment if Python DADA2 is available

# Machine learning for taxonomic classification
scikit-learn>=1.0

# Data processing & pipeline
Click>=8.0

# Optional: Jupyter notebook environment
# jupyter>=1.0
# notebook>=7.0
# ipython>=9.0

# Optional: Development
# pytest>=7.0
# pytest-cov>=4.0
# black>=23.0
